a bug fix plus changes as per featurecats request
