+++ src/main/java/wagner/stephanie/lizzie/gui/LizzieFrame.java
@@ -514,7 +514,7 @@ public class LizzieFrame extends JFrame {
         if (validLastWinrate && validWinrate) {
         	String text;
             if( Lizzie.config.config.getJSONObject("ui").getBoolean("handicap-instead-of-winrate") ) {
-                text=String.format(": %.2f%%", Lizzie.winrateToHandicap(100 - lastWR - curWR));
+            	text=String.format(": %.2f", Lizzie.leelaz.winrateToHandicap(100-curWR) - Lizzie.leelaz.winrateToHandicap(lastWR));
         	} else {
                 text=String.format(": %.1f%%", 100 - lastWR - curWR);
         	}
